# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Road Trip Planner API with AI, Maps, and Music
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: trip
  - name: trips
  - name: ai
  - name: places
  - name: music
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /trip/vehicle-check:
    post:
      operationId: checkVehicle
      tags: [trip]
      summary: Check vehicle suitability
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VehicleCheckRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VehicleCheckResponse"

  /trip/plan:
    post:
      operationId: planTrip
      tags: [trip]
      summary: Plan a road trip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripPlanRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripPlanResponse"

  /trip/adjust:
    post:
      operationId: adjustTrip
      tags: [trip]
      summary: Adjust trip plan due to delay
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripAdjustRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripPlanResponse"

  /trip/traffic:
    post:
      operationId: getTrafficInfo
      tags: [trip]
      summary: Get traffic conditions along route
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrafficRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrafficResponse"

  /trips:
    get:
      operationId: listTrips
      tags: [trips]
      summary: List all saved trips
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  trips:
                    type: array
                    items:
                      $ref: "#/components/schemas/SavedTrip"
                required: [trips]
    post:
      operationId: saveTrip
      tags: [trips]
      summary: Save a trip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveTripRequest"
      responses:
        "201":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedTrip"

  /trips/{id}:
    get:
      operationId: getTrip
      tags: [trips]
      summary: Get a saved trip by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedTrip"
    patch:
      operationId: updateTrip
      tags: [trips]
      summary: Update a saved trip (name, favorite, custom stops)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTripRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedTrip"
    delete:
      operationId: deleteTrip
      tags: [trips]
      summary: Delete a saved trip
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /ai/recommendations:
    post:
      operationId: getAiRecommendations
      tags: [ai]
      summary: Get AI-powered personalized trip recommendations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiRecommendationsRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiRecommendationsResponse"

  /ai/music:
    post:
      operationId: getAiMusicPlaylist
      tags: [music]
      summary: Get AI-curated music playlist for the trip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MusicPlaylistRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MusicPlaylistResponse"

  /places/nearby:
    post:
      operationId: searchNearbyPlaces
      tags: [places]
      summary: Search for nearby places along the route
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NearbyPlacesRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NearbyPlacesResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    VehicleCheckRequest:
      type: object
      properties:
        travelers:
          type: integer
          minimum: 1
        vehicleType:
          type: string
        destination:
          type: string
      required: [travelers, vehicleType]

    VehicleCheckResponse:
      type: object
      properties:
        suitable:
          type: boolean
        suitabilityScore:
          type: integer
        issues:
          type: array
          items:
            type: string
        recommendations:
          type: array
          items:
            $ref: "#/components/schemas/VehicleRecommendation"
        message:
          type: string
      required: [suitable, suitabilityScore, issues, recommendations, message]

    VehicleRecommendation:
      type: object
      properties:
        vehicleType:
          type: string
        reason:
          type: string
        capacity:
          type: integer
        features:
          type: array
          items:
            type: string
      required: [vehicleType, reason, capacity, features]

    TripPlanRequest:
      type: object
      properties:
        travelers:
          type: integer
          minimum: 1
        vehicleType:
          type: string
        startLocation:
          type: string
        destination:
          type: string
        startTime:
          type: string
        preferences:
          type: array
          items:
            type: string
      required: [travelers, vehicleType, startLocation, destination, startTime]

    TripPlanResponse:
      type: object
      properties:
        routes:
          type: array
          items:
            $ref: "#/components/schemas/Route"
        selectedRoute:
          $ref: "#/components/schemas/Route"
        stops:
          type: array
          items:
            $ref: "#/components/schemas/Stop"
        mealRecommendations:
          type: array
          items:
            $ref: "#/components/schemas/MealStop"
        estimatedArrival:
          type: string
        totalDistance:
          type: string
        summary:
          type: string
      required: [routes, stops, mealRecommendations, estimatedArrival, totalDistance, summary]

    Route:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [fastest, scenic, balanced]
        duration:
          type: string
        distance:
          type: string
        highlights:
          type: array
          items:
            type: string
        description:
          type: string
      required: [id, name, type, duration, distance, highlights, description]

    Stop:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [attraction, viewpoint, rest_area, gas_station, town, landmark, custom]
        description:
          type: string
        estimatedArrivalTime:
          type: string
        estimatedStopDuration:
          type: string
        distanceFromStart:
          type: string
        mustVisit:
          type: boolean
        openHours:
          type: string
        isOpen:
          type: boolean
        coordinates:
          $ref: "#/components/schemas/Coordinates"
        isCustom:
          type: boolean
        rating:
          type: number
        photoUrl:
          type: string
      required: [id, name, type, description, estimatedArrivalTime, estimatedStopDuration, distanceFromStart, mustVisit, isOpen]

    MealStop:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        mealType:
          type: string
          enum: [breakfast, lunch, dinner, snack]
        cuisine:
          type: string
        description:
          type: string
        estimatedArrivalTime:
          type: string
        location:
          type: string
        priceRange:
          type: string
          enum: [$, $$, $$$]
        openHours:
          type: string
        isOpen:
          type: boolean
        mustTry:
          type: array
          items:
            type: string
        rating:
          type: number
        alternativeOptions:
          type: array
          items:
            $ref: "#/components/schemas/MealAlternative"
      required: [id, name, mealType, cuisine, description, estimatedArrivalTime, location, priceRange, isOpen, mustTry]

    MealAlternative:
      type: object
      properties:
        name:
          type: string
        cuisine:
          type: string
        priceRange:
          type: string
        openHours:
          type: string
        reason:
          type: string
      required: [name, cuisine, priceRange, reason]

    TripAdjustRequest:
      type: object
      properties:
        originalRequest:
          $ref: "#/components/schemas/TripPlanRequest"
        delayMinutes:
          type: integer
        currentLocation:
          type: string
        reason:
          type: string
      required: [originalRequest, delayMinutes]

    TrafficRequest:
      type: object
      properties:
        startLocation:
          type: string
        destination:
          type: string
        routeType:
          type: string
          enum: [fastest, scenic, balanced]
        departureTime:
          type: string
      required: [startLocation, destination]

    TrafficResponse:
      type: object
      properties:
        conditions:
          type: string
          enum: [clear, light, moderate, heavy, severe]
        delayMinutes:
          type: integer
        incidents:
          type: array
          items:
            $ref: "#/components/schemas/TrafficIncident"
        alternativeRouteAvailable:
          type: boolean
        recommendation:
          type: string
        updatedAt:
          type: string
      required: [conditions, delayMinutes, incidents, alternativeRouteAvailable, recommendation, updatedAt]

    TrafficIncident:
      type: object
      properties:
        type:
          type: string
          enum: [accident, construction, closure, congestion, weather]
        description:
          type: string
        location:
          type: string
        severity:
          type: string
          enum: [low, medium, high]
        delayMinutes:
          type: integer
      required: [type, description, location, severity, delayMinutes]

    SavedTrip:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        travelers:
          type: integer
        vehicleType:
          type: string
        startLocation:
          type: string
        destination:
          type: string
        startTime:
          type: string
        preferences:
          type: array
          items:
            type: string
        plan:
          $ref: "#/components/schemas/TripPlanResponse"
        customStops:
          type: array
          items:
            $ref: "#/components/schemas/Stop"
        isFavorite:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, travelers, vehicleType, startLocation, destination, startTime, isFavorite]

    SaveTripRequest:
      type: object
      properties:
        name:
          type: string
        travelers:
          type: integer
        vehicleType:
          type: string
        startLocation:
          type: string
        destination:
          type: string
        startTime:
          type: string
        preferences:
          type: array
          items:
            type: string
        plan:
          $ref: "#/components/schemas/TripPlanResponse"
        customStops:
          type: array
          items:
            $ref: "#/components/schemas/Stop"
      required: [name, travelers, vehicleType, startLocation, destination, startTime]

    UpdateTripRequest:
      type: object
      properties:
        name:
          type: string
        isFavorite:
          type: boolean
        customStops:
          type: array
          items:
            $ref: "#/components/schemas/Stop"
        plan:
          $ref: "#/components/schemas/TripPlanResponse"

    AiRecommendationsRequest:
      type: object
      properties:
        tripRequest:
          $ref: "#/components/schemas/TripPlanRequest"
        currentPlan:
          $ref: "#/components/schemas/TripPlanResponse"
        userPreferences:
          type: object
          properties:
            interests:
              type: array
              items:
                type: string
            dietaryRestrictions:
              type: array
              items:
                type: string
            budget:
              type: string
              enum: [budget, moderate, luxury]
            travelStyle:
              type: string
              enum: [adventure, relaxed, family, romantic, cultural]
        question:
          type: string
          description: Optional specific question from the user
      required: [tripRequest]

    AiRecommendationsResponse:
      type: object
      properties:
        recommendations:
          type: array
          items:
            $ref: "#/components/schemas/AiRecommendation"
        generalAdvice:
          type: string
        alternativeStops:
          type: array
          items:
            $ref: "#/components/schemas/Stop"
        packingTips:
          type: array
          items:
            type: string
        weatherNote:
          type: string
      required: [recommendations, generalAdvice]

    AiRecommendation:
      type: object
      properties:
        category:
          type: string
          enum: [attraction, restaurant, activity, accommodation, tip]
        name:
          type: string
        description:
          type: string
        reason:
          type: string
        estimatedTime:
          type: string
        location:
          type: string
        priority:
          type: string
          enum: [must_see, recommended, optional]
      required: [category, name, description, reason, priority]

    MusicPlaylistRequest:
      type: object
      properties:
        tripRequest:
          $ref: "#/components/schemas/TripPlanRequest"
        currentTimeOfDay:
          type: string
          enum: [morning, afternoon, evening, night]
        mood:
          type: string
          enum: [energetic, relaxed, adventurous, romantic, focused, fun]
        routeType:
          type: string
          enum: [fastest, scenic, balanced]
        genres:
          type: array
          items:
            type: string
          description: Preferred music genres
      required: [tripRequest, mood]

    MusicPlaylistResponse:
      type: object
      properties:
        playlistName:
          type: string
        playlistDescription:
          type: string
        mood:
          type: string
        tracks:
          type: array
          items:
            $ref: "#/components/schemas/MusicTrack"
        spotifySearchUrl:
          type: string
        totalDuration:
          type: string
        genres:
          type: array
          items:
            type: string
        aiCurationNote:
          type: string
      required: [playlistName, playlistDescription, mood, tracks, genres, aiCurationNote]

    MusicTrack:
      type: object
      properties:
        title:
          type: string
        artist:
          type: string
        album:
          type: string
        duration:
          type: string
        genre:
          type: string
        energy:
          type: string
          enum: [low, medium, high]
        spotifySearchQuery:
          type: string
        youtubeSearchQuery:
          type: string
        reason:
          type: string
          description: Why this song fits the trip
      required: [title, artist, duration, genre, energy, reason]

    NearbyPlacesRequest:
      type: object
      properties:
        location:
          type: string
          description: Location name or "lat,lng" format
        radius:
          type: integer
          description: Search radius in miles
          default: 10
        categories:
          type: array
          items:
            type: string
            enum: [restaurant, gas_station, attraction, rest_area, hotel, coffee, grocery]
        openNow:
          type: boolean
          default: false
        currentTime:
          type: string
          description: ISO 8601 datetime for open-now filtering
      required: [location]

    NearbyPlacesResponse:
      type: object
      properties:
        places:
          type: array
          items:
            $ref: "#/components/schemas/NearbyPlace"
        searchLocation:
          type: string
        totalFound:
          type: integer
      required: [places, searchLocation, totalFound]

    NearbyPlace:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        category:
          type: string
        description:
          type: string
        address:
          type: string
        distance:
          type: string
        openHours:
          type: string
        isOpen:
          type: boolean
        rating:
          type: number
        priceLevel:
          type: integer
          description: 1-4 price level
        coordinates:
          $ref: "#/components/schemas/Coordinates"
        tags:
          type: array
          items:
            type: string
      required: [id, name, category, description, address, distance, isOpen]

    Coordinates:
      type: object
      properties:
        lat:
          type: number
        lng:
          type: number
      required: [lat, lng]