# lib/api-zod/src/generated/types/healthStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Road Trip Planner API with AI, Maps, and Music
 * OpenAPI spec version: 0.2.0
 */

export interface HealthStatus {
  status: string;
}