# lib/api-zod/src/generated/types/mealStop.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Road Trip Planner API with AI, Maps, and Music
 * OpenAPI spec version: 0.2.0
 */
import type { MealAlternative } from "./mealAlternative";
import type { MealStopMealType } from "./mealStopMealType";
import type { MealStopPriceRange } from "./mealStopPriceRange";

export interface MealStop {
  id: string;
  name: string;
  mealType: MealStopMealType;
  cuisine: string;
  description: string;
  estimatedArrivalTime: string;
  location: string;
  priceRange: MealStopPriceRange;
  openHours?: string;
  isOpen: boolean;
  mustTry: string[];
  rating?: number;
  alternativeOptions?: MealAlternative[];
}