# lib/api-zod/src/generated/types/musicPlaylistRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Road Trip Planner API with AI, Maps, and Music
 * OpenAPI spec version: 0.2.0
 */
import type { MusicPlaylistRequestCurrentTimeOfDay } from "./musicPlaylistRequestCurrentTimeOfDay";
import type { MusicPlaylistRequestMood } from "./musicPlaylistRequestMood";
import type { MusicPlaylistRequestRouteType } from "./musicPlaylistRequestRouteType";
import type { TripPlanRequest } from "./tripPlanRequest";

export interface MusicPlaylistRequest {
  tripRequest: TripPlanRequest;
  currentTimeOfDay?: MusicPlaylistRequestCurrentTimeOfDay;
  mood: MusicPlaylistRequestMood;
  routeType?: MusicPlaylistRequestRouteType;
  /** Preferred music genres */
  genres?: string[];
}