# lib/api-zod/src/generated/types/nearbyPlacesRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Road Trip Planner API with AI, Maps, and Music
 * OpenAPI spec version: 0.2.0
 */
import type { NearbyPlacesRequestCategoriesItem } from "./nearbyPlacesRequestCategoriesItem";

export interface NearbyPlacesRequest {
  /** Location name or "lat,lng" format */
  location: string;
  /** Search radius in miles */
  radius?: number;
  categories?: NearbyPlacesRequestCategoriesItem[];
  openNow?: boolean;
  /** ISO 8601 datetime for open-now filtering */
  currentTime?: string;
}