# lib/api-zod/src/generated/types/route.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Road Trip Planner API with AI, Maps, and Music
 * OpenAPI spec version: 0.2.0
 */
import type { RouteType } from "./routeType";

export interface Route {
  id: string;
  name: string;
  type: RouteType;
  duration: string;
  distance: string;
  highlights: string[];
  description: string;
}