# lib/api-zod/src/generated/types/savedTrip.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Road Trip Planner API with AI, Maps, and Music
 * OpenAPI spec version: 0.2.0
 */
import type { Stop } from "./stop";
import type { TripPlanResponse } from "./tripPlanResponse";

export interface SavedTrip {
  id: number;
  name: string;
  travelers: number;
  vehicleType: string;
  startLocation: string;
  destination: string;
  startTime: string;
  preferences?: string[];
  plan?: TripPlanResponse;
  customStops?: Stop[];
  isFavorite: boolean;
  createdAt?: string;
  updatedAt?: string;
}