# lib/api-zod/src/generated/types/stop.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Road Trip Planner API with AI, Maps, and Music
 * OpenAPI spec version: 0.2.0
 */
import type { Coordinates } from "./coordinates";
import type { StopType } from "./stopType";

export interface Stop {
  id: string;
  name: string;
  type: StopType;
  description: string;
  estimatedArrivalTime: string;
  estimatedStopDuration: string;
  distanceFromStart: string;
  mustVisit: boolean;
  openHours?: string;
  isOpen: boolean;
  coordinates?: Coordinates;
  isCustom?: boolean;
  rating?: number;
  photoUrl?: string;
}