# lib/api-zod/src/generated/types/tripPlanResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Road Trip Planner API with AI, Maps, and Music
 * OpenAPI spec version: 0.2.0
 */
import type { MealStop } from "./mealStop";
import type { Route } from "./route";
import type { Stop } from "./stop";

export interface TripPlanResponse {
  routes: Route[];
  selectedRoute?: Route;
  stops: Stop[];
  mealRecommendations: MealStop[];
  estimatedArrival: string;
  totalDistance: string;
  summary: string;
}