# lib/api-zod/src/generated/types/updateTripRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Road Trip Planner API with AI, Maps, and Music
 * OpenAPI spec version: 0.2.0
 */
import type { Stop } from "./stop";
import type { TripPlanResponse } from "./tripPlanResponse";

export interface UpdateTripRequest {
  name?: string;
  isFavorite?: boolean;
  customStops?: Stop[];
  plan?: TripPlanResponse;
}