# lib/api-zod/src/generated/types/vehicleCheckRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Road Trip Planner API with AI, Maps, and Music
 * OpenAPI spec version: 0.2.0
 */

export interface VehicleCheckRequest {
  /** @minimum 1 */
  travelers: number;
  vehicleType: string;
  destination?: string;
}